Comments from Lukas
